+++ packages/backend/src/boot/entry.ts
@@ -14,12 +14,10 @@ import Xev from 'xev';
 import Logger from '@/logger.js';
 import { envOption } from '../env.js';
 import { masterMain } from './master.js';
-import { workerMain } from './worker.js';
+import { parseWorkerArguments, workerMain } from './worker.js';
 import { readyRef } from './ready.js';
 
 import 'reflect-metadata';
-
-process.title = `Misskey (${cluster.isPrimary ? 'master' : 'worker'})`;
 
 Error.stackTraceLimit = Infinity;
 EventEmitter.defaultMaxListeners = 128;
@@ -71,17 +69,21 @@ process.on('exit', code => {
 if (!envOption.disableClustering) {
 	if (cluster.isPrimary) {
 		logger.info(`Start main process... pid: ${process.pid}`);
+		process.title = 'Misskey (master)';
 		await masterMain();
 		ev.mount();
 	} else if (cluster.isWorker) {
 		logger.info(`Start worker process... pid: ${process.pid}`);
-		await workerMain();
+		const workerArguments = parseWorkerArguments(process.env);
+		process.title = `Misskey (worker${workerArguments.__workerName ? `: ${workerArguments.__workerName}` : ''})`;
+		await workerMain(workerArguments);
 	} else {
 		throw new Error('Unknown process type');
 	}
 } else {
 	// 非clusterの場合はMasterのみが起動するため、Workerの処理は行わない(cluster.isWorker === trueの状態でこのブロックに来ることはない)
 	logger.info(`Start main process... pid: ${process.pid}`);
+	process.title = 'Misskey (master)';
 	await masterMain();
 	ev.mount();
 }